fix(status-line): move refresh-gate helper into git-helpers
Keep shouldApplyOpenPrRefresh free of pi-tui imports so Node tests can
cover cwd/branch staleness without loading the status-line module.
